style(App.js): comment out unused Sidebar and SidebarOnDesktop imports fix(Header): add logout button to the header when user is authenticated feat(Header): add isAuth and signUserOut props to Header component feat(CreatePost): add image field to the form validation feat(CreatePost): navigate to home page after adding a post to the database style(Header.module.scss): add styles for logout button The Sidebar and SidebarOnDesktop components are not used in the App.js file, so they are commented out. The Header component now displays a logout button when the user is authenticated. The isAuth and signUserOut props are added to the Header component. The CreatePost component now validates the image field in the form. After adding a post to the database, the user is navigated to the home page. The Header.module.scss file now has styles for the logout button.

+++ src/App.js
@@ -7,8 +7,8 @@ import { signUserAccountOut } from './utils/firebase'
 import Home from './pages/Home'
 import Login from './pages/Login'
 import CreatePost from './pages/CreatePost'
-import Sidebar from './components/SideBar/Sidebar'
-import SidebarOnDesktop from './components/SidebarOnDesktop/SidebarOnDesktop'
+// import Sidebar from './components/SideBar/Sidebar'
+// import SidebarOnDesktop from './components/SidebarOnDesktop/SidebarOnDesktop'
 import Blogpost from './pages/Blogpost'
 import UserInfo from './pages/UserInfo'
 import About from './pages/About'
@@ -18,11 +18,15 @@ function App() {
   const [isAuth, setIsAuth] = useState(localStorage.getItem('isAuth'))
 
   const signUserOut = () => {
-    signUserAccountOut().then(() => {
-      localStorage.clear()
-      setIsAuth(false)
-      window.location.pathname = '/login'
-    })
+    // we must ask for confirmation before logging out
+    // it might happen that the user have clicked logout button by mistake
+    if (window.confirm('Are you sure you want to logout?')) {
+      signUserAccountOut().then(() => {
+        localStorage.clear()
+        setIsAuth(false)
+        window.location.pathname = '/login'
+      })
+    }
   }
   let width = 1000
 
@@ -32,6 +36,8 @@ function App() {
     <Router>
       <Header
         headerShouldBeLarge={['/createpost'].includes(window.location.pathname)}
+        isAuth={isAuth}
+        signUserOut={signUserOut}
       />
 
       {/* previous navigation options are listed below */}
